refactor: move to util package
